Bunch of changes
Switch to Transit to parse JSON, so we immediately get Clojurescript
datastructures, and not #js arrays and objects. This is also a lot
faster than js->clj.

Render the parsed Markdown as HTML. The markdown comes down the API in
Hickory or clojure/data.xml style, and is easy to render. This does
raise the concern that for proper editing and rendering we probably want
to be able to render and parse markdown on the client side. There are JS
libs to do this, but then we would be using a different parser than the
one on the backend, and that's just gonna cause pain. Might have to
round trip to the server for now.

Use figwheel as nREPL and Ring server. One process == win. Pretty cool
how well this works.

Use Sablono instead of om.dom